+++ src/main/java/com/group6/wishshare/web/controller/LoginController.java
@@ -1,10 +1,10 @@
 package com.group6.wishshare.web.controller;
 
-import com.group6.wishshare.data.repository.UserRepositoryImpl;
+import com.group6.wishshare.data.repository.mysql.UserRepositoryImpl;
 import com.group6.wishshare.domain.model.User;
 import com.group6.wishshare.domain.model.type.Gender;
 import com.group6.wishshare.domain.service.LoginException;
-import com.group6.wishshare.domain.service.LoginService;
+import com.group6.wishshare.domain.service.UserService;
 import com.group6.wishshare.web.util.SessionObject;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
@@ -18,7 +18,9 @@ import java.util.Objects;
 
 @Controller
 public class LoginController {
-  private final LoginService LOGIN_CONTROLLER = new LoginService(new UserRepositoryImpl());
+
+
+  private final UserService USER_SERVICE = new UserService(new UserRepositoryImpl());
 
   @GetMapping("/")
   public String getIndex() {
@@ -41,7 +43,7 @@ public class LoginController {
       String email = webRequest.getParameter("email");
       String password = webRequest.getParameter("password");
 
-      User user = LOGIN_CONTROLLER.login(email, password);
+      User user = USER_SERVICE.login(email, password);
       new SessionObject(webRequest).setUser(user);
       return "redirect:dashboard";
     } catch (LoginException e) {
@@ -63,7 +65,7 @@ public class LoginController {
     try {
       if (validatePassword(password1, password2)) {
         User user =
-            LOGIN_CONTROLLER.createUser(
+            USER_SERVICE.createUser(
                 firstname,
                 lastname,
                 LocalDate.parse(Objects.requireNonNull(birthdate)),
