+++ src/main/java/com/group6/wishshare/web/controller/WishlistController.java
@@ -1,7 +1,7 @@
 package com.group6.wishshare.web.controller;
 
-import com.group6.wishshare.data.repository.UserRepositoryImpl;
-import com.group6.wishshare.data.repository.WishlistRepositoryImpl;
+import com.group6.wishshare.data.repository.mysql.UserRepositoryImpl;
+import com.group6.wishshare.data.repository.mysql.WishlistRepositoryImpl;
 import com.group6.wishshare.domain.model.User;
 import com.group6.wishshare.domain.model.Wishlist;
 import com.group6.wishshare.domain.service.UserService;
